feat: support editor actions scaffolds
