+++ taier-ui/src/extensions/action/quickRunSQLAction.ts
@@ -1,6 +1,6 @@
-import { CREATE_MODEL_TYPE, ID_COLLECTIONS } from '@/constant';
+import { ID_COLLECTIONS } from '@/constant';
 import type { CatalogueDataProps, IOfflineTaskProps } from '@/interface';
-import { editorActionBarService } from '@/services';
+import taskRenderService from '@/services/taskRenderService';
 import { isTaskTab } from '@/utils/is';
 import { runTask } from '@/utils/extensions';
 import molecule from '@dtinsight/molecule';
@@ -36,9 +36,9 @@ export default class QuickRunSQLAction extends Action2 {
 		const { current } = molecule.editor.getState();
 		if (current && isTaskTab(current.tab?.id)) {
 			const currentTabData: CatalogueDataProps & IOfflineTaskProps = current?.tab?.data;
-			const taskToolbar = editorActionBarService.getActionBar(
+			const taskToolbar = taskRenderService.renderEditorActions(
 				currentTabData.taskType,
-				currentTabData.createModel === CREATE_MODEL_TYPE.GUIDE,
+				currentTabData,
 			);
 			// 只要当前任务存在运行按钮才可以执行运行命令
 			if (taskToolbar.find((t) => t.id === ID_COLLECTIONS.TASK_RUN_ID)) {
