+++ taier-ui/src/extensions/action/quickSaveTaskAction.ts
@@ -1,14 +1,14 @@
+import { message } from 'antd';
+import { KeyMod, KeyCode } from 'monaco-editor';
 import api from '@/api';
-import { CREATE_MODEL_TYPE, ID_COLLECTIONS } from '@/constant';
-import { CatalogueDataProps, IOfflineTaskProps } from '@/interface';
-import { editorActionBarService } from '@/services';
+import { ID_COLLECTIONS } from '@/constant';
+import taskRenderService from '@/services/taskRenderService';
 import { isTaskTab } from '@/utils/is';
 import saveTask from '@/utils/saveTask';
 import molecule from '@dtinsight/molecule';
 import { Action2 } from '@dtinsight/molecule/esm/monaco/action';
 import { KeybindingWeight } from '@dtinsight/molecule/esm/monaco/common';
-import { message } from 'antd';
-import { KeyMod, KeyCode } from 'monaco-editor';
+import type { CatalogueDataProps, IOfflineTaskProps } from '@/interface';
 
 export default class QuickSaveTaskAction extends Action2 {
 	static readonly ID = 'SaveTask';
@@ -38,9 +38,9 @@ export default class QuickSaveTaskAction extends Action2 {
 		const { current } = molecule.editor.getState();
 		if (current && isTaskTab(current.tab?.id)) {
 			const currentTabData: CatalogueDataProps & IOfflineTaskProps = current?.tab?.data;
-			const taskToolbar = editorActionBarService.getActionBar(
+			const taskToolbar = taskRenderService.renderEditorActions(
 				currentTabData.taskType,
-				currentTabData.createModel === CREATE_MODEL_TYPE.GUIDE,
+				currentTabData,
 			);
 			if (taskToolbar.find((t) => t.id === ID_COLLECTIONS.TASK_SAVE_ID)) {
 				saveTask()
